Server/requirements.txt
accelerate==0.33.0
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
bcrypt==4.2.0
boto3==1.35.11
botocore==1.35.11
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
diffusers==0.29.2
ecdsa==0.19.0
fastapi==0.112.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauth==1.0.1
packaging==24.1
passlib==1.7.4
pillow==10.4.0
psutil==6.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
sympy==1.13.1
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
zipp==3.19.2
